+++ lib/src/Model/reader_result.dart
@@ -3,7 +3,7 @@ import 'package:epubease/src/Model/last_place_model.dart';
 class ReaderResult {
   final LastPlaceModel lactPlace;
   final List<LastPlaceModel> chapters;
-  final int totalProgress;
+  final double totalProgress;
 
   ReaderResult({
     required this.lactPlace,
